+++ RogueLike/Assets/Armas/Melee.asset
@@ -14,5 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   name: Melee
   icon: {fileID: -1851786855, guid: 1f51656b056b6ae43b6e400da16b3bb1, type: 3}
-  damage: 25
-  price: 0
+  damage: 5
