+++ RogueLike/Assets/Armas/Sniper.asset
@@ -14,5 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   name: Sniper
   icon: {fileID: 1120435292, guid: 24770f737b71a3f4e99f6a0089885131, type: 3}
-  damage: 100
-  price: 100
+  damage: 0
