+++ RogueLike/Assets/InputSystem/Inputs.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 157977c3f7db4cb49ac101d86c97801d
+guid: 758b831c039b1fa40ae0a3e76d9d0592
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
