--- RogueLike/Assets/Inputs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12c2c729cc9f75e46bb65437c6abeed1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
