+++ RogueLike/Assets/Prefab/TIENDA/SlotTienda.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 62a3cca3c39f09142a413f8cf7c01b1e
-folderAsset: yes
-DefaultImporter:
+guid: cd811ce98aeed7e4986a72c1a5311c75
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
