--- RogueLike/Assets/Prefab/TIENDA/Slott.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e074c26cb75b1084daf04a2496750366
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
