+++ RogueLike/Assets/Scripts/Arma.cs
@@ -6,7 +6,6 @@ public class Arma : ScriptableObject
     new public string name = "arma";
     public Sprite icon = null;
     public float damage = 0f;
-    public int price = 0;
 
     public virtual void Use()
     {
