--- RogueLike/Assets/Scripts/CANVAS/INVENTARIO.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 09a0ce0a8e93a3b4283af9a03641d848
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
