--- RogueLike/Assets/Scripts/CANVAS/INVENTARIO/Inventario.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6552fc9ba7a857f42ae30a2225abd286
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
