--- RogueLike/Assets/Scripts/CANVAS/TIENDA.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0b7d37d43a515b1429e414ac3e18bbec
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
