--- RogueLike/Assets/Scripts/CANVAS/TIENDA/Tienda.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-
-public class Tienda : MonoBehaviour
-{
-    public static Tienda instance;
-
-    private void Awake()
-    {
-        if (instance == null)
-        {
-            DontDestroyOnLoad(gameObject);
-            instance = this;
-        }
-        else Destroy(gameObject);
-    }
-}
