--- RogueLike/Assets/Scripts/CANVAS/TIENDA/Tienda.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0032a5559eea2e142a605d9067f2fe64
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
