--- RogueLike/Assets/Scripts/CANVAS/TIENDA/TiendaUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d5dc96be98a9699469fa378ac417f5b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
