+++ RogueLike/Assets/Scripts/GameManager.cs
@@ -1,43 +1,19 @@
 using System.Collections;
 using System.Collections.Generic;
-using UnityEditor.SearchService;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 public class GameManager : MonoBehaviour
 {
-    public static GameManager instance;
-    Inventario inventario;
-
-    private void Awake()
-    {
-        if (instance == null)
-        {
-            DontDestroyOnLoad(gameObject);
-            instance = this;
-        }
-        else Destroy(gameObject);
-    }
     public int run = 1;
     // Start is called before the first frame update
     void Start()
     {
-        inventario = Inventario.instance;
+        
     }
 
     // Update is called once per frame
     void Update()
     {
         
-    }
-    public bool openTienda()
-    {
-        if (SceneManager.GetActiveScene().name == "Lobby") return true;
-        else return false;
-    }
-    public bool buyWeapon(Arma arma)
-    {
-        if (!inventario.items.Contains(arma) && arma.price>0) return true;
-        return false;
     }
 }
