+++ RogueLike/Assets/Scripts/Inventario.cs
@@ -4,26 +4,20 @@ using UnityEngine;
 
 public class Inventario : MonoBehaviour
 {
+    public static Inventario instance;
+
+    private void Awake()
+    {
+        if (instance != null) return;
+        instance = this; 
+    }
+
     public delegate void OnItemChanged();
     public OnItemChanged onItemChangedCallback;
 
     public int space = 6;
-    public static Inventario instance;
-
-    private void Start()
-    {
-        DontDestroyOnLoad(gameObject);
-    }
-    private void Awake()
-    {
-        if (instance == null)
-        {
-            DontDestroyOnLoad(gameObject);
-            instance = this;
-        }
-        else Destroy(gameObject);
-    }
     public List<Arma> items = new List<Arma>();
+
     public void Add(Arma arma)
     {
         if (items.Count < space)
@@ -32,4 +26,5 @@ public class Inventario : MonoBehaviour
         }
         if (onItemChangedCallback != null) onItemChangedCallback.Invoke();
     }
+
 }
