+++ RogueLike/Assets/Scripts/Inventario.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 350292f6561ae5a4e92fafd43453d534
+guid: 9fd03b91e941a184488bb1b2935c7056
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
