+++ RogueLike/Assets/Scripts/InventarioSlot.cs.meta
