+++ RogueLike/Assets/TextMesh Pro/Examples & Extras/Scripts/Benchmark02.cs
@@ -71,6 +71,7 @@ namespace TMPro.Examples
                     // Canvas WorldSpace Camera
                     GameObject go = new GameObject();
                     Canvas canvas = go.AddComponent<Canvas>();
+                    canvas.worldCamera = Camera.main;
 
                     go.transform.localScale = new Vector3(0.1f, 0.1f, 0.1f);
                     go.transform.position = new Vector3(Random.Range(-95f, 95f), 5f, Random.Range(-95f, 95f));
